+++ Source/Moq.Analyzers.Test/CallbackSignatureShouldMatchMockedMethodAnalyzerTests.ShouldFailWhenIncorrectCallbacks.verified.txt
@@ -1,6 +1,6 @@
 ﻿Diagnostic 1
 	Id: Moq1100
-	Location: SourceFile(Test0.cs[403..410))
+	Location: SourceFile(Test0.cs[420..427))
 	Highlight: (int i)
 	Lines: mock.Setup(x => x.Do(It.IsAny<string>())).Callback((int i) => { });
 	Severity: Warning
@@ -8,7 +8,7 @@
 
 Diagnostic 2
 	Id: Moq1100
-	Location: SourceFile(Test0.cs[479..501))
+	Location: SourceFile(Test0.cs[496..518))
 	Highlight: (string s1, string s2)
 	Lines: mock.Setup(x => x.Do(It.IsAny<string>())).Callback((string s1, string s2) => { });
 	Severity: Warning
@@ -16,7 +16,7 @@ Diagnostic 2
 
 Diagnostic 3
 	Id: Moq1100
-	Location: SourceFile(Test0.cs[609..628))
+	Location: SourceFile(Test0.cs[626..645))
 	Highlight: (string s1, int i1)
 	Lines: mock.Setup(x => x.Do(It.IsAny<int>(), It.IsAny<string>(), It.IsAny<DateTime>())).Callback((string s1, int i1) => { });
 	Severity: Warning
@@ -24,7 +24,7 @@ Diagnostic 3
 
 Diagnostic 4
 	Id: Moq1100
-	Location: SourceFile(Test0.cs[703..710))
+	Location: SourceFile(Test0.cs[720..727))
 	Highlight: (int i)
 	Lines: mock.Setup(x => x.Do(It.IsAny<List<string>>())).Callback((int i) => { });
 	Severity: Warning
