+++ Source/Moq.Analyzers.Test/CallbackSignatureShouldMatchMockedMethodAnalyzerTests.ShouldPassWhenCorrectSetupAndCallbacks.verified.txt
@@ -1,0 +1,1 @@
+﻿emptyString
