+++ Source/Moq.Analyzers.Test/CallbackSignatureShouldMatchMockedMethodAnalyzerTests.ShouldPassWhenCorrectSetupAndParameterlessCallbacks.verified.txt
@@ -1,0 +1,1 @@
+﻿emptyString
