+++ Source/Moq.Analyzers.Test/CallbackSignatureShouldMatchMockedMethodAnalyzerTests.ShouldPassWhenCorrectSetupAndReturns.verified.txt
@@ -1,0 +1,1 @@
+﻿emptyString
