+++ Source/Moq.Analyzers.Test/CallbackSignatureShouldMatchMockedMethodAnalyzerTests.ShouldPassWhenCorrectSetupAndReturnsAndCallbacks.verified.txt
@@ -1,0 +1,1 @@
+﻿emptyString
