+++ Source/Moq.Analyzers.Test/CallbackSignatureShouldMatchMockedMethodCodeFixTests.cs
@@ -10,9 +10,33 @@ namespace Moq.Analyzers.Test;
 public class CallbackSignatureShouldMatchMockedMethodCodeFixTests : CodeFixVerifier
 {
     [Fact]
-    public Task ShouldSuggestQuickFixIfBadParameters()
+    public Task ShouldPassWhenCorrectSetupAndReturns()
     {
-        return Verify(VerifyCSharpFix(File.ReadAllText("Data/CallbackSignatureShouldMatchMockedMethod.cs")));
+        return Verify(VerifyCSharpFix(File.ReadAllText("Data/CallbackSignatureShouldMatchMockedMethod.MyGoodSetupAndReturns.cs")));
+    }
+
+    [Fact]
+    public Task ShouldSuggestQuickFixWhenIncorrectCallbacks()
+    {
+        return Verify(VerifyCSharpFix(File.ReadAllText("Data/CallbackSignatureShouldMatchMockedMethod.TestBadCallbacks.cs")));
+    }
+
+    [Fact]
+    public Task ShouldPassWhenCorrectSetupAndCallbacks()
+    {
+        return Verify(VerifyCSharpFix(File.ReadAllText("Data/CallbackSignatureShouldMatchMockedMethod.TestGoodSetupAndCallback.cs")));
+    }
+
+    [Fact]
+    public Task ShouldPassWhenCorrectSetupAndParameterlessCallbacks()
+    {
+        return Verify(VerifyCSharpFix(File.ReadAllText("Data/CallbackSignatureShouldMatchMockedMethod.TestGoodSetupAndParameterlessCallback.cs")));
+    }
+
+    [Fact]
+    public Task ShouldPassWhenCorrectSetupAndReturnsAndCallbacks()
+    {
+        return Verify(VerifyCSharpFix(File.ReadAllText("Data/CallbackSignatureShouldMatchMockedMethod.TestGoodSetupAndReturnsAndCallback.cs")));
     }
 
     protected override CodeFixProvider GetCSharpCodeFixProvider()
